backend/requirements.txt
anyio==4.12.0
asgiref==3.9.2
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
bleach==6.3.0
boto3==1.42.31
botocore==1.42.31
brotli==1.2.0
CacheControl==0.14.4
cachetools==6.2.4
cbor2==5.8.0
certifi==2025.8.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
Django==5.2.6
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-dotenv==1.4.2
django-filter==25.2
django-grappelli==4.0.3
django-jasmine==0.5.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.17
django-nested-admin==4.1.6
django-rest-framework==0.1.0
django-rosetta==0.10.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
drf-yasg==1.21.11
firebase_admin==7.1.0
fonttools==4.61.1
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
nltk==3.9.1
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polib==1.2.0
proto-plus==1.27.0
protobuf==6.33.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-monkey-business==1.1.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rosetta==0.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scipy==1.16.2
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==67.0
webencodings==0.5.1
zope.interface==8.2
zopfli==0.4.0
